Separates membergroup moderation actions from non-mod actions
They were horribly entangled!
